third_party/WebKit/LayoutTests/fast/xmlhttprequest/xmlhttprequest-set-responsetype-expected.txt

CONSOLE WARNING: The provided value 'dkjdfkjdfkj' is not a valid enum value of type XMLHttpRequestResponseType.
CONSOLE WARNING: The provided value 'asdfasdfasd' is not a valid enum value of type XMLHttpRequestResponseType.
This test checks if setting non-supported type does not throw an exception

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseType is ""
PASS xhr.responseType is "arraybuffer"
PASS xhr.responseType is "blob"
PASS xhr.responseType is "document"
PASS xhr.responseType is "json"
PASS xhr.responseType is "text"
PASS xhr.responseType = "dkjdfkjdfkj" did not throw exception.
PASS xhr.responseType = "text"; xhr.responseType = "asdfasdfasd"; xhr.responseType is "text"
PASS successfullyParsed is true

TEST COMPLETE
